Fix dtype, model navigation, and mask shape issues for working DDTree
- Cast attention mask to model dtype (bfloat16)
- Use dflash-mlx's _target_text_model and _lm_head_logits for correct model navigation
- Read actual KV cache offset instead of assuming it equals start
- Cast draft logits to float32 before NumPy conversion

First benchmark: DDTree-16 achieves 1.14x speedup over DFlash at 1K tokens.

+++ ddtree_mlx/runtime.py
@@ -167,7 +167,7 @@ def generate_ddtree_once(
 
         # --- TREE BUILD ---
         build_start = time.perf_counter_ns()
-        draft_logits_np = np.array(draft_logits[0], copy=False)
+        draft_logits_np = np.array(draft_logits[0].astype(mx.float32), copy=False)
         tree = build_ddtree_tree(draft_logits_np, budget=tree_budget)
         root_token = int(staged_first.item())
         compiled = compile_tree(tree, root_token, prefix_len=start)
